sync

+++ trainers/trainer.py
@@ -255,9 +255,9 @@ class Trainer:
                     print("INPUT_IDS: ", batch['input_ids'])
                     print("device: ", self.client.device)
                     print("dtype, shape: ", batch['input_ids'].dtype, batch['input_ids'].shape)
-                    input_ids = batch['input_ids'].to(model_dtype).to(self.client.device)
-                    label_ids = batch['labels'].to(model_dtype).to(self.client.device)
-                    attention_mask=batch['attention_mask'].to(model_dtype).to(self.client.device)
+                    input_ids = batch['input_ids'].to(self.client.device)
+                    label_ids = batch['labels'].to(self.client.device)
+                    attention_mask=batch['attention_mask'].to(self.client.device)
                     output_ids = self.client.model.generate(
                         input_ids=input_ids,
                         attention_mask=attention_mask,
@@ -309,9 +309,9 @@ class Trainer:
         with torch.no_grad():
             for batch in self.client.eval_loader_genr:
                 try:
-                    input_ids = batch['input_ids'].to(model_dtype).to(self.client.device)
-                    label_ids = batch['labels'].to(model_dtype).to(self.client.device)
-                    attention_mask=batch['attention_mask'].to(model_dtype).to(self.client.device)
+                    input_ids = batch['input_ids'].to(self.client.device)
+                    label_ids = batch['labels'].to(self.client.device)
+                    attention_mask=batch['attention_mask'].to(self.client.device)
                     output_ids = self.client.model.generate(
                         input_ids=input_ids,
                         attention_mask=attention_mask,
